requirements.txt
pytesseract
tesseract
openai
streamlit
Pillow